requirements.txt
asgiref==3.8.1
boto3==1.40.6
boto3-stubs==1.41.1
botocore==1.40.6
botocore-stubs==1.42.5
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2
django-debug-toolbar==6.1.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
idna==3.10
jmespath==1.0.1
mysqlclient==2.2.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==7.1.0
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
types-awscrt==0.29.2
types-s3transfer==0.16.0
tzdata==2025.2
urllib3==2.5.0
